backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
motor>=3.3.2
beanie>=1.23.6
python-dotenv>=1.0.0
google-generativeai>=0.3.2
boto3>=1.29.7
pydantic>=2.10.0,<3.0.0
